Remove component tests for now
