+++ src/NServiceBus.Core.Tests/NServiceBus.Core.Tests.csproj
@@ -111,37 +111,6 @@
     <Compile Include="AssemblyScanner\When_scanning_top_level_only.cs" />
     <Compile Include="Audit\AuditConfigReaderTests.cs" />
     <Compile Include="Causation\AttachCausationHeadersBehaviorTests.cs" />
-    <Compile Include="ComponentTests\OutboxStorageTests.cs" />
-    <Compile Include="ComponentTests\PersistenceTestsConfiguration.cs" />
-    <Compile Include="ComponentTests\Sagas\AnotherSagaWithCorrelatedPropertyData.cs" />
-    <Compile Include="ComponentTests\Sagas\AnotherSagaWithoutCorrelationProperty.cs" />
-    <Compile Include="ComponentTests\Sagas\SagaPersisterTests.cs" />
-    <Compile Include="ComponentTests\Sagas\SagaWithComplexType.cs" />
-    <Compile Include="ComponentTests\Sagas\SagaWithComplexTypeEntity.cs" />
-    <Compile Include="ComponentTests\Sagas\SagaWithCorrelationPropertyData.cs" />
-    <Compile Include="ComponentTests\Sagas\SagaWithoutCorrelationPropertyData.cs" />
-    <Compile Include="ComponentTests\Sagas\SimpleSagaEntitySaga.cs" />
-    <Compile Include="ComponentTests\Sagas\TestSagaData.cs" />
-    <Compile Include="ComponentTests\Sagas\When_completing_a_saga_loaded_by_id.cs" />
-    <Compile Include="ComponentTests\Sagas\When_completing_a_saga_with_correlation_property.cs" />
-    <Compile Include="ComponentTests\Sagas\When_completing_a_saga_with_no_defined_correlation_property.cs" />
-    <Compile Include="ComponentTests\Sagas\When_multiple_workers_retrieve_same_saga_on_different_threads.cs" />
-    <Compile Include="ComponentTests\Sagas\When_multiple_workers_retrieve_same_saga_on_the_same_thread.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_a_saga_with_an_escalated_DTC_transaction.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_a_saga_with_complex_types.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_a_saga_with_no_defined_unique_property.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_a_saga_with_the_same_unique_property_as_another_saga.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_a_saga_with_the_same_unique_property_as_a_completed_saga.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_different_sagas_with_no_defined_unique_properties.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_different_sagas_with_unique_properties.cs" />
-    <Compile Include="ComponentTests\Sagas\When_persisting_the_same_saga_twice_in_two_sessions_on_the_same_thread.cs" />
-    <Compile Include="ComponentTests\Sagas\When_retrieving_the_same_saga_twice.cs" />
-    <Compile Include="ComponentTests\Sagas\When_saga_not_found_return_default.cs" />
-    <Compile Include="ComponentTests\Sagas\When_updating_a_saga_with_no_defined_unique_property.cs" />
-    <Compile Include="ComponentTests\Sagas\When_updating_a_saga_with_the_same_unique_property_value.cs" />
-    <Compile Include="ComponentTests\Sagas\When_worker_tries_to_complete_saga_update_by_another.cs" />
-    <Compile Include="ComponentTests\SubscriptionStorageTests.cs" />
-    <Compile Include="ComponentTests\TimeoutStorageTests.cs" />
     <Compile Include="Config\When_using_initialization.cs" />
     <Compile Include="ContextBagTests.cs" />
     <Compile Include="Correlation\CorrelationContextExtensionsTests.cs" />
@@ -173,7 +142,6 @@
     <Compile Include="Msmq\MsmqAddressTests.cs" />
     <Compile Include="Msmq\MsmqSubscriptionStorageTests.cs" />
     <Compile Include="Performance\MessageDurability\DetermineMessageDurabilityBehaviorTests.cs" />
-    <Compile Include="PersistenceTestsConfiguration.cs" />
     <Compile Include="Persistence\InMemory\InMemorySubscriptionStorageTests.cs" />
     <Compile Include="Persistence\InMemory\SagaWithoutUniquePropertyData.cs" />
     <Compile Include="Persistence\InMemory\When_persisting_a_saga_with_complex_types.cs" />
