+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -139,6 +139,7 @@
     <Compile Include="Persistence\Development\SagaPersister\DevelopmentSyncronizedStorage.cs" />
     <Compile Include="Persistence\Development\SagaPersister\DevelopmentSyncronizedStorageSession.cs" />
     <Compile Include="Persistence\Development\SagaPersister\SagaManifest.cs" />
+    <Compile Include="Persistence\Development\SagaPersister\SagaManifestCollection.cs" />
     <Compile Include="Persistence\Development\SagaPersister\SagaStorageFile.cs" />
     <Compile Include="Persistence\Msmq\SubscriptionStorage\MsmqSubscriptionMessage.cs" />
     <Compile Include="Persistence\Msmq\SubscriptionStorage\MsmqSubscriptionStorageConfigurationExtensions.cs" />
