+++ src/NServiceBus.Core/Persistence/Development/SagaPersister/DevelopmentSagaPersistence.cs
@@ -1,9 +1,7 @@
 ﻿namespace NServiceBus.Features
 {
     using System;
-    using System.Collections.Generic;
     using System.IO;
-    using System.Runtime.Serialization.Json;
     using System.Threading.Tasks;
     using Extensibility;
     using NServiceBus.Outbox;
@@ -28,26 +26,8 @@
             var storageLocation = context.Settings.Get<string>(StorageLocationKey);
 
             var allSagas = context.Settings.Get<SagaMetadataCollection>();
-            var sagaManifests = new Dictionary<Type, SagaManifest>();
-
-            foreach (var metadata in allSagas)
-            {
-                var sagaStorageDir = Path.Combine(storageLocation, metadata.SagaType.FullName.Replace("+", ""));
-
-                if (!Directory.Exists(sagaStorageDir))
-                {
-                    Directory.CreateDirectory(sagaStorageDir);
-                }
-
-                var manifest = new SagaManifest
-                {
-                    StorageDirectory = sagaStorageDir,
-                    Serializer = new DataContractJsonSerializer(metadata.SagaEntityType),
-                    SagaEntityType = metadata.SagaEntityType
-                };
-
-                sagaManifests[metadata.SagaEntityType] = manifest;
-            }
+            
+            var sagaManifests = new SagaManifestCollection(allSagas,storageLocation);
 
             context.Container.ConfigureComponent<DevelopmentSyncronizedStorage>(DependencyLifecycle.SingleInstance);
             context.Container.ConfigureComponent<DevelopmentStorageAdapter>(DependencyLifecycle.SingleInstance);
